+++ vidarr-server/src/main/java/ca/on/oicr/gsi/vidarr/server/Main.java
@@ -125,7 +125,6 @@ import java.util.concurrent.Semaphore;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.function.Function;
-import java.util.regex.Matcher;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.xml.stream.XMLStreamException;
@@ -1968,7 +1967,7 @@ public final class Main implements ServerConfig {
                                 .apply(
                                     new ExtractInputVidarrIds(
                                         MAPPER, workflowRun.getArguments().get(param.getKey()))))
-                    .map(Main::hashFromAnalysisId)
+                    .map(DatabaseBackedProcessor::hashFromAnalysisId)
                     .collect(Collectors.toCollection(TreeSet::new)),
                 workflowRun.getExternalKeys());
 
@@ -2122,13 +2121,6 @@ public final class Main implements ServerConfig {
       exchange.getResponseHeaders().put(Headers.CONTENT_TYPE, CONTENT_TYPE_TEXT);
       exchange.getResponseSender().send(e.getMessage());
     }
-  }
-
-  public static String hashFromAnalysisId(String id) {
-    Matcher matcher = BaseProcessor.ANALYSIS_RECORD_ID.matcher(id);
-    if (!matcher.matches())
-      throw new IllegalStateException("Failed to match ANALYSIS_RECORD_ID regex to id: " + id);
-    return matcher.group("hash");
   }
 
   private void loadDataIntoDatabase(
